+++ src/FromState.ts
@@ -3,10 +3,10 @@
  *
  * @since 2.11.0
  */
-import type { Chain, Chain2, Chain3, Chain4 } from './Chain'
-import type { Endomorphism } from './Endomorphism'
+import { Chain, Chain2, Chain3, Chain4 } from './Chain'
+import { Endomorphism } from './Endomorphism'
 import { flow } from './function'
-import type { HKT2, Kind2, Kind3, Kind4, URIS2, URIS3, URIS4 } from './HKT'
+import { HKT2, Kind2, Kind3, Kind4, URIS2, URIS3, URIS4 } from './HKT'
 import * as S from './State'
 
 import State = S.State