+++ src/function.ts
@@ -55,7 +55,8 @@ export const getSemigroup = <S>(S: Semigroup<S>) => <A = never>(): Semigroup<(a:
  * Unary functions form a monoid as long as you can provide a monoid for the codomain.
  *
  * @example
- * import { Predicate, getMonoid } from 'fp-ts/Predicate'
+ * import { Predicate } from 'fp-ts/Predicate'
+ * import { getMonoid } from 'fp-ts/function'
  * import * as B from 'fp-ts/boolean'
  *
  * const f: Predicate<number> = (n) => n <= 2