+++ test/ReaderEither.ts
@@ -272,4 +272,15 @@ describe('ReaderEither', () => {
     U.deepStrictEqual(f(_.right(1))({}), 'right')
     U.deepStrictEqual(f(_.left('a'))({}), 'left')
   })
+
+  it('fromReaderK', () => {
+    const ma = _.fromReaderK((n: number): R.Reader<number, number> => (c) => n * c)
+    U.deepStrictEqual(ma(3)(2), E.right(6))
+  })
+
+  it('chainReaderK', () => {
+    const f = _.chainReaderK((n: number): R.Reader<number, number> => (c) => n * c)
+    U.deepStrictEqual(pipe(_.right(3), f)(2), E.right(6))
+    U.deepStrictEqual(pipe(_.left('a'), f)(2), E.left('a'))
+  })
 })