+++ CHANGELOG.md
@@ -49,6 +49,7 @@ high state of flux, you're at risk of it changing without notice.
   - add `FromState` module
   - add `FromThese` module
   - add `void` module
+  - add `FromReader` module
   - `Array`
     - add `prependW`, `appendW` (@thewilkybarkid)
     - add `fromOption`, `fromPredicate` (@cdimitroulas)